+++ changelog.md
@@ -1,3 +1,9 @@
+# **v25-05-18.p1**
+
+## Additions
+- Visual themes which change your background and button colors.
+
+
 # **v25-08-15.2**
 
 ## Additions